% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formats.R
\docType{data}
\name{formats}
\alias{formats}
\title{Formats des fichiers VID-HOSP}
\format{Un data.frame avec 9 variables
\itemize{
\item \emph{nom} Nom de la variable tel que décrit dans le manuel de l'ATIH
\item \emph{taille} Taille du champ
\item \emph{debut} Position du premier caractère de la variable. NA pour les champs à position relative.
\item \emph{fin} Position du dernier caractère de la variable. NA pour les champs à position relative.
\item \emph{obligatoire} O = obligatoire, F = Facultatif. NA et N ne sont pas documentés.
\item \emph{nom_variable} Nom de la variable à utiliser dans R. Est une normalisation à 20 caractères avec uniquement des caractères alphanumériques minuscules ASCII et des underscores à la place des espaces.
\item \emph{format_version} Version du format
\item \emph{position_variable} Indique si le champ est à position variable. TRUE si \emph{debut} et \emph{fin} sont NA.
}}
\source{
\itemize{
\item Formats 2018 : \url{https://www.atih.sante.fr/formats-pmsi-2018}
\item Formats 2017 : \url{https://www.atih.sante.fr/formats-pmsi-2017}
}
}
\usage{
formats
}
\description{
Formats permettant d'importer les fichiers VID-HSOP. Ils sont importés directement depuis les PDF fournis par l'ATIH à l'aide de \href{https://tabula.technology/}{tabula} qui produit des CSV puis de scripts adaptés pour chaque années disponibles dans `inst/extdata``
}
\keyword{datasets}
