% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_vidhosp.R
\name{importe_vidhosp}
\alias{importe_vidhosp}
\title{Import un fichier VID-HOSP}
\usage{
importe_vidhosp(chemin_vidhosp, partie_variable = FALSE,
  etiquettes = TRUE)
}
\arguments{
\item{chemin_vidhosp}{Chemin d'accès vers le ficheir VID-HOSP}

\item{partie_variable}{Si FALSE, n'importe pas la partie variable. Non fonctionnel pour l'instant, partie variable à implémenter}

\item{etiquettes}{Ajouter les étiquettes de variables pour plus facilement identifier. A implémenter.}
}
\value{
Un tibble avec les données du VID-HOSP
}
\description{
Lit le fichier VID-HOSP, détermine son format, récupère les positions des variables et l'importe.
}
