% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importe_vidhosp.R
\name{lis_format}
\alias{lis_format}
\title{Lit format d'un VID-HOSP}
\usage{
lis_format(chemin_vidhosp, debut = 49, fin = 52)
}
\arguments{
\item{chemin_vidhosp}{Chemin d'accès vers le ficheir VID-HOSP}

\item{debut}{Position du premier caractère}

\item{fin}{Position du dernier caractère}
}
\value{
Une chaine de caractère avec la version du format
}
\description{
Lit le format d'un fichier VID-HOSP. Récupère la première ligne du fichier et lit le format qui est normalement toujours à la même position
}
