% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_format_vidhosp.R
\name{normalise}
\alias{normalise}
\title{Normalise noms}
\usage{
normalise(x, limite = 26)
}
\arguments{
\item{x}{chaine de caractère correspondant aux noms des colonnes}

\item{limite}{nombre de caractère maximum pour le nom de la colonne dans R}
}
\value{
Un vecteur de caractères normalisés
}
\description{
Simplifie les description des variables pour en faire des noms de variables
adaptées à la manipulation dans R. Fonction d'aide pour importer les formats. Utilisé pour importer les formats.
}
\seealso{
Dans le package, le répertoire \code{/inst/extdata}.
}
